Code Clean up

+++ StudyN/Views/AddCategoryPage.xaml.cs
@@ -52,15 +52,15 @@ namespace StudyN.Views
 			if (GlobalAppointmentData.EditCategory == null) {
 				// creates the category
 				cat = GlobalAppointmentData.CalendarManager.CreateCategory(this.name.Text, this.color, 
-																			((double)this.colorPicker.PointerRingPositionXUnits), 
-																			((double)this.colorPicker.PointerRingPositionYUnits));
+																			this.colorPicker.PointerRingPositionXUnits, 
+																			this.colorPicker.PointerRingPositionYUnits);
 			}
 			else
 			{
 				// edit the category
 				GlobalAppointmentData.CalendarManager.EditCategory(this.name.Text, this.color, 
-																	((double)this.colorPicker.PointerRingPositionXUnits), 
-																	((double)this.colorPicker.PointerRingPositionYUnits),
+																	this.colorPicker.PointerRingPositionXUnits, 
+																	this.colorPicker.PointerRingPositionYUnits,
 																	GlobalAppointmentData.EditCategory.Id);
 				cat = GlobalAppointmentData.EditCategory;
 				GlobalAppointmentData.EditCategory = null;
